+++ perf/__main__.py
@@ -85,6 +85,8 @@ def create_parser():
                              help='Absolute minimum of speed in percent to '
                                   'consider that a benchmark is significant '
                                   '(default: 0%%)')
+            cmd.add_argument('--table', action="store_true",
+                             help='Render a table')
         input_filenames(cmd)
 
     # stats
@@ -347,6 +349,15 @@ def _display_common_metadata(metadatas, lines):
 
 def cmd_compare(args):
     from perf._compare import compare_suites
+
+    if getattr(args, 'table', False):
+        if args.group_by_speed:
+            print("--table doesn't support --group-by-speed yet")
+            sys.exit(1)
+
+        if args.min_speed:
+            print("--table doesn't support --min-speed yet")
+            sys.exit(1)
 
     data = load_benchmarks(args)
     if data.get_nsuite() < 2: