+++ doc/changelog.rst
@@ -1,18 +1,28 @@
 Changelog
 =========
 
+Version 0.9.4
+-------------
+
+New features:
+
+* Add ``--compare-to`` option to the :ref:`Runner CLI <runner_cli>`
+* :ref:`compare_to <compare_cmd>` command: Add ``--table`` option to render a table
+
+Bugfixes:
+
+* Fix the ``abs_executable()`` function used to find the absolute path to the
+  Python program. Don't follow symbolic links to support correctly virtual
+  environments.
+
 Version 0.9.3 (2017-01-16)
 --------------------------
 
-* :ref:`compare_to <compare_cmd>` command: Add ``--table`` option to render a table
 * Fix the Windows support.
 * system: Don't try to read or write CPU frequency when the
   /sys/devices/system/cpu/cpu0/cpufreq/ directory doesn't exist. For example,
   virtual machines don't have this directory.
 * Fix a ``ResourceWarning`` in ``BenchmarkSuite.dump()`` for gzip files.
-* Fix the abs_executable() function used to find the absolute path to the
-  Python program. Don't follow symbolic links to support correctly virtual
-  environments.
 
 Version 0.9.2 (2016-12-15)
 --------------------------